Update the site to new design
This is the third site design (iirc). The main goal was to remove my dependence on client-side scripts to render content. The result, hopefully, is a site that should function almost identically with and without JavaScript. The site itself also trims out a lot of old content (sorry magic 8 ball). If you want to retrieve these things they're still part of the commit history so don't worry.