Modify theme to improve mobile margins.
Previously, on very thin devices the margins and text alignment would
stick to the right side of the screen. On thin (i.e., less than or equal
to 8.5 inches wide) screens margins are now added to the <main> element
to give text some breathing room. Similarly, text is not justified on
thin screens but aligned with the default setting (probably left).